+++ app/api/views.py
@@ -16,7 +16,7 @@ from django.contrib.auth.decorators import login_required, permission_required
 from drf_spectacular.utils import extend_schema, extend_schema_view, OpenApiResponse, OpenApiParameter
 from rest_framework import status
 from rest_framework import viewsets
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, action
 from rest_framework.response import Response
 from host.object_store import ObjectStore
 from host.models import Aperture
@@ -47,8 +47,19 @@ from host.log import get_logger
 logger = get_logger(__name__)
 
 
+def stream_download_file(file_path):
+    # Stream the data file from the S3 bucket
+    s3 = ObjectStore()
+    object_key = os.path.join(settings.S3_BASE_PATH, file_path.strip('/'))
+    filename = os.path.basename(file_path)
+    obj_stream = s3.stream_object(object_key)
+    response = StreamingHttpResponse(streaming_content=obj_stream)
+    response["Content-Disposition"] = f"attachment; filename={filename}"
+    return response
+
 ############################################################
 # Filter Sets
+
 
 class TransientFilter(django_filters.FilterSet):
     redshift_lte = django_filters.NumberFilter(
@@ -158,6 +169,11 @@ class CutoutViewSet(viewsets.ReadOnlyModelViewSet):
     serializer_class = CutoutSerializer
     filter_backends = (DjangoFilterBackend,)
     filterset_class = CutoutFilter
+
+    @action(methods=['get'], detail=True, url_path="download")
+    def download(self, request, pk=None):
+        cutout = self.get_object()
+        return stream_download_file(cutout.fits.name)
 
 
 class FilterViewSet(viewsets.ReadOnlyModelViewSet):